+++ trn_invadjfile2.php
@@ -11,7 +11,9 @@ require "includes/main_header.php";
 //     $password   = $rs_salesfile1['password'];
 //     // loop here
 // }
+date_default_timezone_set('Asia/Manila');
 
+$header_usercode = '';
 
 if(isset($_POST['recid_hidden']) && !empty($_POST['recid_hidden'])){
     $select_db_docnum1='SELECT * FROM tranfile1 WHERE recid=?';
@@ -48,6 +50,7 @@ if(isset($_POST['recid_hidden']) && !empty($_POST['recid_hidden'])){
         $payment_details  = $rs_docnum1['paydetails'];
         $remarks  = $rs_docnum1['remarks'];
         $ordernum  = $rs_docnum1['ordernum'];
+        $header_usercode = isset($rs_docnum1['usercode']) ? trim((string)$rs_docnum1['usercode']) : '';
 
     }
 }else{
@@ -61,8 +64,6 @@ if(isset($_POST['recid_hidden']) && !empty($_POST['recid_hidden'])){
         $docnum  = "ADJ-00001";
     }
 
-    date_default_timezone_set('Asia/Manila');
-
     $trndate = date('m/d/Y');
     $trntot = '';
     $customer_name = '';
@@ -72,6 +73,59 @@ if(isset($_POST['recid_hidden']) && !empty($_POST['recid_hidden'])){
     $payment_details  = '';
     $remarks  = '';
     $ordernum = '';
+}
+
+$session_usercode = '';
+if(isset($_SESSION['usercode']) && trim((string)$_SESSION['usercode']) !== ''){
+    $session_usercode = trim((string)$_SESSION['usercode']);
+}else if(isset($_POST["usercode_hidden"]) && trim((string)$_POST["usercode_hidden"]) !== ''){
+    $session_usercode = trim((string)$_POST["usercode_hidden"]);
+}
+
+$display_usercode = ($session_usercode !== '') ? $session_usercode : $header_usercode;
+$display_userdesc = '';
+if($display_usercode !== ''){
+    $select_user = "SELECT userdesc FROM users WHERE usercode = ? LIMIT 1";
+    $stmt_user = $link->prepare($select_user);
+    $stmt_user->execute(array($display_usercode));
+    $rs_user = $stmt_user->fetch();
+    if(!empty($rs_user) && isset($rs_user['userdesc'])){
+        $display_userdesc = $rs_user['userdesc'];
+    }
+}
+
+$warehouse_options = array();
+$stmt_warehouse = $link->prepare("SELECT warcde, warehouse_name FROM warehouse ORDER BY warehouse_name ASC");
+$stmt_warehouse->execute();
+while($rs_warehouse = $stmt_warehouse->fetch()){
+    $warehouse_options[] = array(
+        'warcde' => $rs_warehouse['warcde'],
+        'warehouse_name' => $rs_warehouse['warehouse_name']
+    );
+}
+
+$warehouse_floor_map = array();
+$stmt_floor = $link->prepare("SELECT warehouse_floor_id, warcde, floor_no FROM warehouse_floor ORDER BY floor_no ASC, floor_name ASC, warehouse_floor_id ASC");
+$stmt_floor->execute();
+while($rs_floor = $stmt_floor->fetch()){
+    $floor_warcde = isset($rs_floor['warcde']) ? (string)$rs_floor['warcde'] : '';
+    if(!isset($warehouse_floor_map[$floor_warcde])){
+        $warehouse_floor_map[$floor_warcde] = array();
+    }
+    $warehouse_floor_map[$floor_warcde][] = array(
+        'warehouse_floor_id' => $rs_floor['warehouse_floor_id'],
+        'floor_no' => $rs_floor['floor_no']
+    );
+}
+
+$warehouse_staff_options = array();
+$stmt_staff = $link->prepare("SELECT warehouse_staff_id, fname, lname FROM warehouse_staff ORDER BY fname ASC, lname ASC");
+$stmt_staff->execute();
+while($rs_staff = $stmt_staff->fetch()){
+    $warehouse_staff_options[] = array(
+        'warehouse_staff_id' => $rs_staff['warehouse_staff_id'],
+        'staff_name' => trim($rs_staff['fname'].' '.$rs_staff['lname'])
+    );
 }
 
 
@@ -210,6 +264,18 @@ if(isset($_POST['recid_hidden']) && !empty($_POST['recid_hidden'])){
                                         </td>                                  
                                     </tr>
 
+                                    <tr class="m-1 edit_row salesfile1" style="border-bottom:3px solid #cccccc ">
+                                        <td colspan="3">
+                                            <div class="row px-2">
+                                                <div class="p-3">
+                                                    <label for="userdesc_display" style="font-weight:bold">User:</label>
+                                                    <input type="text" class="form-control" name="userdesc_display" id="userdesc_display" value="<?php echo htmlspecialchars($display_userdesc, ENT_QUOTES); ?>" readonly>
+                                                    <input type="hidden" name="usercode_1" id="usercode_1" value="<?php echo htmlspecialchars($display_usercode, ENT_QUOTES); ?>">
+                                                </div>
+                                            </div>
+                                        </td>
+                                    </tr>
+
                                     <tr class="m-1 salesfile1" id="tr_access_data">
                                         <td id="main_chk_div" colspan="3">
                                             <div class='w-100 d-flex justify-content-center'>
@@ -311,6 +377,39 @@ if(isset($_POST['recid_hidden']) && !empty($_POST['recid_hidden'])){
                                 </div>
                             </div>
 
+                            <div class="row m-3">
+                                <div class="col-12">
+                                    <label for="warcde_add">Warehouse</label>
+                                    <select name="warcde_add" id="warcde_add" class="form-select">
+                                        <option value="">Select Warehouse</option>
+                                        <?php foreach($warehouse_options as $warehouse_option): ?>
+                                            <option value="<?php echo htmlspecialchars($warehouse_option['warcde'], ENT_QUOTES); ?>"><?php echo htmlspecialchars($warehouse_option['warehouse_name'], ENT_QUOTES); ?></option>
+                                        <?php endforeach; ?>
+                                    </select>
+                                </div>
+                            </div>
+
+                            <div class="row m-3">
+                                <div class="col-12">
+                                    <label for="warehouse_floor_id_add">Warehouse Floor</label>
+                                    <select name="warehouse_floor_id_add" id="warehouse_floor_id_add" class="form-select">
+                                        <option value="">Select Warehouse Floor</option>
+                                    </select>
+                                </div>
+                            </div>
+
+                            <div class="row m-3">
+                                <div class="col-12">
+                                    <label for="warehouse_staff_id_add">Warehouse Staff</label>
+                                    <select name="warehouse_staff_id_add" id="warehouse_staff_id_add" class="form-select">
+                                        <option value="">Select Warehouse Staff</option>
+                                        <?php foreach($warehouse_staff_options as $staff_option): ?>
+                                            <option value="<?php echo htmlspecialchars($staff_option['warehouse_staff_id'], ENT_QUOTES); ?>"><?php echo htmlspecialchars($staff_option['staff_name'], ENT_QUOTES); ?></option>
+                                        <?php endforeach; ?>
+                                    </select>
+                                </div>
+                            </div>
+
 
                             <input type="hidden" name="itmcde_add_hidden" id="itmcde_add_hidden">
                             <div class="row m-2">
@@ -365,6 +464,39 @@ if(isset($_POST['recid_hidden']) && !empty($_POST['recid_hidden'])){
                                 <div class="col-12">
                                     <label for="">Amount</label>
                                     <input type="text" name="amount_edit" id="amount_edit" class="form-control" autocomplete="off" readonly>
+                                </div>
+                            </div>
+
+                            <div class="row m-3">
+                                <div class="col-12">
+                                    <label for="warcde_edit">Warehouse</label>
+                                    <select name="warcde_edit" id="warcde_edit" class="form-select">
+                                        <option value="">Select Warehouse</option>
+                                        <?php foreach($warehouse_options as $warehouse_option): ?>
+                                            <option value="<?php echo htmlspecialchars($warehouse_option['warcde'], ENT_QUOTES); ?>"><?php echo htmlspecialchars($warehouse_option['warehouse_name'], ENT_QUOTES); ?></option>
+                                        <?php endforeach; ?>
+                                    </select>
+                                </div>
+                            </div>
+
+                            <div class="row m-3">
+                                <div class="col-12">
+                                    <label for="warehouse_floor_id_edit">Warehouse Floor</label>
+                                    <select name="warehouse_floor_id_edit" id="warehouse_floor_id_edit" class="form-select">
+                                        <option value="">Select Warehouse Floor</option>
+                                    </select>
+                                </div>
+                            </div>
+
+                            <div class="row m-3">
+                                <div class="col-12">
+                                    <label for="warehouse_staff_id_edit">Warehouse Staff</label>
+                                    <select name="warehouse_staff_id_edit" id="warehouse_staff_id_edit" class="form-select">
+                                        <option value="">Select Warehouse Staff</option>
+                                        <?php foreach($warehouse_staff_options as $staff_option): ?>
+                                            <option value="<?php echo htmlspecialchars($staff_option['warehouse_staff_id'], ENT_QUOTES); ?>"><?php echo htmlspecialchars($staff_option['staff_name'], ENT_QUOTES); ?></option>
+                                        <?php endforeach; ?>
+                                    </select>
                                 </div>
                             </div>
 
@@ -443,6 +575,25 @@ if(isset($_POST['recid_hidden']) && !empty($_POST['recid_hidden'])){
         <script>
 
         var trncde = $("#trncde_hidden").val();
+        var warehouseFloorMap = <?php echo json_encode($warehouse_floor_map); ?>;
+
+        function rebuildFloorOptions(selectId, warcde, selectedFloorId){
+            var $select = $("#" + selectId);
+            if($select.length === 0){
+                return;
+            }
+
+            var options = "<option value=''>Select Warehouse Floor</option>";
+            var floors = warehouseFloorMap[warcde] || [];
+
+            for(var i = 0; i < floors.length; i++){
+                var floor = floors[i];
+                var selected = (selectedFloorId && selectedFloorId === floor.warehouse_floor_id) ? " selected" : "";
+                options += "<option value='" + floor.warehouse_floor_id + "'" + selected + ">" + floor.floor_no + "</option>";
+            }
+
+            $select.html(options);
+        }
 
         $(document).ready(function(){
             var docnum = $("#docnum_hidden").val();
@@ -458,6 +609,14 @@ if(isset($_POST['recid_hidden']) && !empty($_POST['recid_hidden'])){
                     }
 
             })
+
+            $("#warcde_add").on("change", function(){
+                rebuildFloorOptions("warehouse_floor_id_add", $(this).val(), "");
+            });
+
+            $("#warcde_edit").on("change", function(){
+                rebuildFloorOptions("warehouse_floor_id_edit", $(this).val(), "");
+            });
         });
 
         function calcTotal(event){
@@ -723,6 +882,9 @@ if(isset($_POST['recid_hidden']) && !empty($_POST['recid_hidden'])){
                     $("#itmqty_add").val('');
                     $("#itmcde_add").val('');
                     $("#itmcde_add_hidden").val('');
+                    $("#warcde_add").val('');
+                    $("#warehouse_staff_id_add").val('');
+                    rebuildFloorOptions("warehouse_floor_id_add", "", "");
                     $(".error_msg_add_modal").html('');
                     $("#insert_modal_sales").modal("show");
 
@@ -855,6 +1017,9 @@ if(isset($_POST['recid_hidden']) && !empty($_POST['recid_hidden'])){
                             $("#price_edit").val(xdata["retEdit"]["untprc"]);
                             $("#amount_edit").val(xdata["retEdit"]["extprc"]);
                             $("#itmqty_edit").val(xdata["retEdit"]["itmqty"]);
+                            $("#warcde_edit").val(xdata["retEdit"]["warcde"]);
+                            rebuildFloorOptions("warehouse_floor_id_edit", xdata["retEdit"]["warcde"], xdata["retEdit"]["warehouse_floor_id"]);
+                            $("#warehouse_staff_id_edit").val(xdata["retEdit"]["warehouse_staff_id"]);
                             // $("#itmcde_edit option[text=" + xdata["retEdit"]["itmdsc"] +"]").prop("selected", true);
                             // $("#itmcde_edit option:contains("+xdata["retEdit"]["itmdsc"]+")").prop("selected", true);
                             // var optionsThatContainValue = $("#itmcde_edit").find('option').filter(function() {
@@ -889,4 +1054,3 @@ if(isset($_POST['recid_hidden']) && !empty($_POST['recid_hidden'])){
 <?php
     require "includes/main_footer.php";
 ?>
-